Reliability changs to open. Allows passing transport options to Ranch.
This commit changes shotgun:open to optionally accept a map
containing Ranch transport options (keyed to transport_opts)
and/or a gun:await_up timeout (keyed to timeout). The default
values are [] and 5000, respectively.

shotgun:open now *blocks* until either gun establishes a connection,
or the user-specified timeout. This makes it more difficult to
send a request to gun before gun has established a connection with
the remote host.

If the timeout is reached, the FSM will be stopped with reason
'gun_open_timeout'. If gun fails while attempting to establish a connection,
the FSM will be stopped with reason 'gun_open_failed'. See the comments near
the code that handles this stuff for some somewhat important information.

In addition to the changes surrounding the use of gun:await_up, one can now
pass options through to the underlying Ranch transport. This allows you to
-for instance- bind a particular shotgun connection to a particular interface
on a multi-homed machine.

Existing users of open/2 or open/3 will not be broken by this change.