Cancel work when request times out.
This addresses Issue #109, and then some.

When a request is made, shotgun:request makes a fresh reference
and passes that along with the work to the shotgun FSM.
When a request times out, shotgun:request now sends an async
request to cancel the now-useless unit of work, passing along
the reference for identification, before notifying the client
of the request timeout.

Things to know:

* Gun has its own request queue. So, if a request has made it
  to gun, cancellation of the request will only squelch future
  messages that would have come from that request. It is
  -however- safe to bounce back to at_rest (as we now do), as
  messages from the cancelled request will not be sent back to
  the FSM.
* Removal of queued requests makes use of queue:filter. For large
  queues, this is kinda not fast.

Minor changes:

* state typedef moved up with the other typedefs.
* enqueue_work_or_stop slightly simplified.