Maybe fixes Issue #95
Some spelunking leads me to believe that commit
0243a27304faf0c58eff74103757eaa2229a37b7
accidentally replaced a queue append operation
with an assignment of the return value of
gen_fsm:reply/2, (whose return value you are
supposed to ignore.). This value was then assigned
to the responses key in the FSM's state map,
which was only *supposed* to contain data of
type queue:queue().

This commit restores the queue append operation,
and *appears* to address Issue #95.